Merge pull request #1971 from RosettaCommons/jadolfbr/fix_append_pose_by_jump
This fixes a bug where energies were not being correctly cleared when appending the pose by a jump.  Most other append/delete functions are calling this clear properly.  

This led to an issue where the concatenated pose was not able to be dumped properly.  
Found by @lqtza and confirmed as missing in the function by @roccomoretti 

The two functions where we now properly clear the energies are:
 - append_pose_by_jump
 - delete_residue_range_slow

Old repository SHA1: df69e27715af57d2484324b102d88d235a3e923a