Merge pull request #1990 from RosettaCommons/roccomoretti/fix_AntibodyDatabaseManager
AntibodyDatabaseManager: extend temporaries until after we're done with the query.

In cppdb::statement::bind(), the passed value should not be a temporary return value from a function. If it is, it will be destroyed before the query completes, leading to havoc and bad results. You need to extend the lifetime of the returned value by assigning it to a variable so it doesn't get destroyed before you're done with the query.

Old repository SHA1: ad756b0ec2d129de3785c242ab31d9e7466c1780