Merge pull request #1983 from RosettaCommons/roccomoretti/update_sqlite
Update SQLite code to the most recent version: 3.16.2

The version of SQLite we had in external (3.7.4) has a number of issues which are detected by the Undefined Behavior Sanitizer (ubsan) tests. Taking a closer look, this looked to definitely be a deficiency in SQLite 3.7.4, rather than how we were using it. Updating to the most recent SQLite version appears to fix these issues.

Some small changes were needed in the SandwichFeatures code to properly order results -- SQL does not require that results be in any particular order, but SandwichFeatures was taking advantage of the fact that SQLite 3.7.4 (but not 3.16.2) ordered the results in a consistent manner. 

Old repository SHA1: 04df72ad40cc7487f92c9cfe940daeaeedc19899