Merge pull request #2013 from RosettaCommons/GerardDaniel/scoreTermValueBasedSelector
This is a new residue selector that selects residues that score within the threshold limits of a score term of choice. 
A subset of residues can be given as input by specifying residue number, selector sub tag or previously defined selector. By default it works on ALL residues
If a score function is not given, the default score function will be used. In apply() a copy of the pose is scored.
If "total" is given as the score term to be used, all weighted per residue scores are summed up and the value compared to the threshold.
A few regression test seem to fail due to insignificant changes in atom coordinates reported.
Soon, integration test will be added by Jason Klima and Benjamin Basanta.

Old repository SHA1: f0ca84ed378e2d5b0b739b89d3540831d6e9b744