Merge pull request #2041 from RosettaCommons/jadolfbr/ccd_ends_small_loops
This special cases small loops so that they are not included in the CCD arms in GraftMovers located in protocols/grafting.  For example, a loop that is 4 residues that you are inserting with 2 residue flexibility on either side will fail.  This changes the insert flexibility where needed so things won't fail. A warning is given for 1 residue insertions. 

Old repository SHA1: 5bc1a1400ba3e159f38f41ee7f4ce65fcb0f9398