Merge pull request #2074 from RosettaCommons/vmullig/fix_christine_nmethylation_issue
Make the UnsatSelector compatible with symmetry

Christine Kang reported an issue in trying to N-methylate peptides with symmetry.  It turns out that the problem isn't with the N-methylation code, but with the UnsatSelector, which was being used to select positions in which the amide proton isn't involved in a hydrogen bond.  This pull request aims to make the UnsatSelector compatible with symmetry.  While I'm at it, I'm adding serialization support for the UnsatSelector, too.

Tasks:
- [x] Make the UnsatSelector symmetry-compatible.
- [x] Make the UnsatSelector serializable.
- [x] Beauty.

While I'm at it:
- [x] Modify FastRelax so that the scorefunction is made symmetric or asymmetric to match the pose.  It's silly to ask the user to keep declaring separate symmetric or asymmetric scorefunctions -- it's just one more barrier to learning Rosetta.
- [x] Do the same with FastDesign.

Old repository SHA1: ffb98a7e9006fd365dc3961731fe01766c5243a5