LabelPoseFromResidueSelectorMover and SplitAndMixPoseMover
SplitAndMixPoseMover. By default extracts all the pieces of a pose as selected through a ResidueSelector but it also allows to then mix the order in which the new pose contains the different pieces. Furthermore it allows to merge pieces coming from different chains into one chain. Because it comes from a ResidueSelector the pieces are, logically disconnected. thus, it produces a multi-chain pose or a pose with chain breaks. Rewrites the PDBInfo.
 LabelPoseFromResidueSelector. Adds or removes labels on residues selected through a ResidueSelector. Thought to work in combination with ResiduePDBInfoHasLabel.

----
An additional 5 commit messages were squashed into this commit:
--
LabelPoseFromResidueSelectorMover: quick fix of a couple of bugs.
--
Allow LabelPoseFromResidueSelectorMover to also remove lables with the reverse option.
--
LabelPoseFromResidueSelectorMover. Provided a ResidueSelector, this will assign a given label (property) to the selected residues.
To be used in pair with the ResiduePDBInfoHasLabel.
--
Fix small issue in SecondaryStructureSelector so that the minE, minH will actually apply.
Fix issue with the default value for the gdt function of RmsdFromResidueSelector.
Added SplitAndMixPoseMover. By default extracts all the pieces of a pose as selected through a ResidueSelector but it also allows to then mix the order in which the new pose contains the different pieces.
--
Fixed RmsdFromResidueSelector. Calculates rmsd throught the RmsdEvaluator
--

Old repository SHA1: 33e71d50908555c9dfb210c62f1745c5687e18bb