Merge pull request #2112 from RosettaCommons/aleaverfay/beautify_Mar2017
Aleaverfay/beautify mar2017

We resolved at the recent WRC meeting that we would be beautifying the code on a per-pull-request basis, with a test that will fail if any of your source files have not been beautified. To get that working, master will have to be beautified again. To that end, I've updated the beautifier to handle some aspects of C++ that it couldn't before (lambda functions with semicolons, macros that can't be followed with a semicolon; namespace aliasing) so that now all of master beautifies.

This is a pre-beautification version of master that fixes things that needed fixin' so that the beautifier will work on it. I'll perform the actual beautification once Sergey has the testing server ready to run the new test.

Old repository SHA1: f6f13d701802fd47a9a4791780e0388f0ff6881e