Merge pull request #2250 from RosettaCommons/vmullig/tweak_perturb_bundle
Add options to PerturbBundle mover to allow parameter values to be set directly.

This allows the PerturbBundle mover to set the absolute value of a parameter if the user so wishes, instead of perturbing a parameter value.

Tasks:
- [x] Add the options.
- [x] Update `parse_my_tag()`.
- [x] Update `provide_xml_schema()`.
- [x] Update `apply()`.
- [x] Integration test.
- [x] Beauty.
- [x] Documentation.

Also:
- [x] Fix current problems with `provide_xml_schema()` function.

Note that the new integration test is set to use the current default scorefunction, be that talaris2014 or ref2015.  This test shouldn't change significantly (except for the score line in the output PDB file) with the switch to ref2015, and I want it to continue to use whatever the current default Rosetta scorefunction is in the future.

Old repository SHA1: ba0f510132c092e0f33ce6922016ed415d0877b6