Merge pull request #2299 from RosettaCommons/vmullig/add_trimesic_acid
Re-activiate commented-out TMA unit test

The added unit test produces weird NaN behaviour ONLY in the Mac/clang build.  I think that this is illustrating an underlying bug in the way that the Mac/clang build is doing math in `core::kinematics`, and this might be related to the existing unit test failure for that build (where different numbers are appearing than what we're getting for the other builds).  It's easiest to debug the problem if the master branch is highlighting it.  We might get to this at the noncanonicals meeting this week.

Old repository SHA1: c33afffc5a88f5d9af785c0f517d42458885727f