Removed assumed connections between non-polymers and glycans.
There was a problem when a non-polymeric HETATM (metals e.t.c.) appears before a protein-bound glycan in the pdb file.
There was a check whether something is NOT a lower terminus, but also the previous residue is not a polymer.
If that is true, Rosetta makes that residue a terminus, because a bond non-polymer -> polymer doesn't make sense.

I added another condition, which checks if residue i has a lower_connect to another polymer to get around that.

Old repository SHA1: 7880b6f57ecb65d69e9c8faf3321246e7e625de2