Merge pull request #2373 from RosettaCommons/vmullig/fix_genkic_sugars
Fix the genkic_sugars integration test in debug mode

Some debug_assert is tripping up the test. I need to figure out what's wrong and fix it.

Update: the logic for detecting noncanonical connections assumed that the only time that a polymer residue would lack a lower connection would be if it were a lower terminus type. This assumption is not true for all glycans. The simple fix is to check for a lower connection instead of checking for the lower terminus type.

Old repository SHA1: e21a430f963c7556cc3c6b1dda89da1ed2be3c83