Merge pull request #2492 from RosettaCommons/revert-2383-bfrenz/auto_glycan_with_links_records
Revert "Bfrenz/auto glycan with links records"

Reverts RosettaCommons/main#2383

This PR causes most of our straight-from-the-pdb benchmark structures to fail loading and breaks the logic of the option in the first place. @raemisch and I will work with @BrandonFrenz in order to get this into Rosetta in a way that does not break existing functionality or logic. In the interim, we will add a set of structures that need to load either as unit tests or an integration test now that we actually have them. Either way they would fail loading.

Old repository SHA1: 9314ed8edc325ca310a9393c6a29cfec788a0536