Merge pull request #2487 from RosettaCommons/jklai/hydrate
This PR adds in the default weights for score12 with the Hydrate/SPaDES protocol (other scoring function weights to come in the future). Also made a few fixes and optimizations that were identified by Andy Watkins. Thanks to him for identifying the issues and helping me with fixing them. The hydrate integration tests are expected to change. The other tests are not expected to change except for a few tests that had previously failed (clang code quality).

Old repository SHA1: a1e6d752093a3e1b40681be0094319e1a74706e0