Merge pull request #2510 from RosettaCommons/sboyken/updated_holes
Adding options to HolesFilter to only calculate holes score for a residue selection.  The holes calculation is performed on the Pose as whole (ignoring the ResidueSelector), but when the time comes to report the score, only the atoms in the residue selector are summed.  The Holes score is a sum of individual atoms/residues anyway (technically "PoseBalls"), so by only reporting a specific selection, we should be able to get location-specific data. 

Old repository SHA1: 18d0b408df4c08a6554887e9a1b7d05a270f3963