Merge pull request #2556 from RosettaCommons/roccomoretti/xml_to_MMfactory
Change parse_my_tag() to use MoveMapFactories.

The parse_movemap() functionality used in parse_my_tag() took a Pose. This was less than desirable, as there's no guarantee that the pose that the move map will be used on will be related in any significant way to the input pose being passed to parse_my_tag().

Now the parse_my_tag() will generate a core::select::movemap::MoveMapFactory object from the tag specification, rather than a MoveMap itself. This will be then applied against the pose in the apply() function to get a MoveMap that is actually used. By waiting for apply() (and re-generating the MoveMap for each call to apply), the MoveMap thus generated will reflect the status of the pose as it is then, rather than the pose as it was in the input.

Old repository SHA1: 75460266bdf98bbd346c114c7e07653c487a1d80