Merge pull request #2575 from RosettaCommons/aleaverfay/jd3_output_structure_naming
JD3: output indices change

The JobOutputIndex is a 4-tuple of integers representing a primary- and a secondary-index
as well as an upper bound on the number of primary- and secondary-indices.

The interface to the PoseOutputters changes slightly.

Derived job queens of the SJQ may override the assign_output_index method to give the just-about-to-be-output Poses different indices if she wants. This is particularly necessary for multi-stage jobs, where there might be more than one InnerLarvalJob (and so the nstruct index is not unique!).

This pull request is a launch point for PR #2613 which further changes the way the JobQueen interacts with generating output. The changes in this PR are not overridden by PR #2613, but if you are writing a new JQ and this PR is disruptive, then you will possibly find the next one disruptive too. Please email me if this affects you.

Old repository SHA1: a31cc2c37c0dd9fff7bf41a78975ef1cba1a6477